Feat: refactor block color swatch to match other content blocks
