+++ src/js/components/ContentBlocks/BlockColorSwatch/BlockColorSwatch.js
@@ -1,7 +1,7 @@
 /* @flow */
 import React from 'react';
 import Box from 'grommet/components/Box';
-import ColorTypeList from '../Shared/ColorTypeList';
+import ColorTypeList from './internal/ColorTypeList';
 
 export default function BlockCarouselSwatch(props: {
   color?: {
