+++ src/js/components/ContentBlocks/BlockColorSwatch/index.js
@@ -1,3 +1,3 @@
-export BlockColorSwatch from './Component';
-export BlockColorSwatchForm from './Form';
-export BlockColorSwatchWireframe from './Wireframe';
+export BlockColorSwatch from './BlockColorSwatch';
+export BlockColorSwatchForm from './BlockColorSwatchForm';
+export BlockColorSwatchWireframe from './BlockColorSwatchWireframe';
