+++ src/js/components/ContentBlocks/BlockColorSwatch/internal/ColorTypeList.js
